feat(tauri-docs): define DocEntry interface and refactor docSearch to use it
