+++ src/extensions/tauri-docs/docSearch.ts
@@ -1,14 +1,6 @@
 import Fuse from "fuse.js";
 import { tauriDocEntries } from "./docEntries";
-
-// Define the structure for documentation entries
-export interface DocEntry {
-  title: string;
-  url: string;
-  description: string;
-  keywords: string[];
-  category: string;
-}
+import type { DocEntry } from "./types/DocEntry";
 
 // Configure Fuse.js for fuzzy searching
 const fuseOptions = {
