+++ src/extensions/tauri-docs/types/DocEntry.ts
@@ -1,0 +1,8 @@
+// Define the structure for documentation entries
+export interface DocEntry {
+  title: string;
+  url: string;
+  description: string;
+  keywords: string[];
+  category: string;
+}
